requirements.txt
langchain-google-genai
streamlit
load-dotenv
langchain
google-generativeai